All methods from flume handler distributed into api directory. All namespaces return using hello functions.
